requirements.txt
streamlit>=1.39.0
pillow>=10.3.0
numpy>=1.26.0
opencv-python-headless>=4.10.0.82
requests>=2.32.0
python-dotenv>=1.0.1
